NEIC-30: Updates for ubuntu:latest, python 3.10, werkzeug 2.3.7
Due to 441 reported CVEs in this image (Ubuntu focal packages and
python libraries), I found it necessary to update the base OS to
ubuntu:latest. Doing so brings in Python 3.10 and updated pipenv. To
resolve the python library CVEs, update the dependencies which include
werkzeug 2.3.7 which necessitates a few minor changes to the
application.